Create FixedLayout map
This will be the base map for the game, in the future will try making
something more interesting (ProcGen, maybe?),
but for now this is good enough.
About half of it is done,
commiting now because some editor fuckery was messing with the map;

Added RiderLink plugin; had to add it to the project
because adding it to the engine was causing the build to fail
due to some bug;

Added Powershell script I had around to help cleanup the project
when stuff breaks.