Implement stalking behavior
Using Behavior Tree for simple chasing logic;
already integrated with Qulock component to stop chasing
when the player is looking at the enemy.

Setting up the collision box revealed some problems with the
"bounds" approach for computing support vertexes
and detecting an actor's presence within the frustum,
but it will work for now.